\name{abremLoglike}
\alias{abremLoglike}

\title{Log likelihood for weibull and lognormal distributions}

\description{
 This function generates the log-likelihood for the chosen distribution having suitable parameters from input data such as required by mlefit.
}

\usage{
abremLoglike(x, par, dist="weibull" )
}

\arguments{
  \item{x}{A dataframe such as generated by mleframe with column names 'left', 'right' and optionally 'qty'.
    Exact failure data (occurrences) have same time entered in both 'left' and 'right' columns.
	Suspension data has last known life[time] entered in 'left' column and -1 entered in 'right' column.
	The left(early) interval bound for left-censored data must be entered as zero. (NA is not accepted).
  }
  \item{par}{A vector of parameters taken in the same order as the underlying R distribution.
  That is, for weibull the order of parameters is (shape, scale).  For lognormal (meanlog, sdlog).
  (Unsuitable parameters cannot be pre-tested, but may cause the function to fail with NaN, or inf values.)
  }
  \item{dist}{A string defining a distribution to be fit. Implemented distributions are "weibull" (default) and "lognormal".
  (Only 2-parameter models are accepted.)
  }
}

\value{
  Returns a scalar vector holding the value of log-likelihood.  Unsuitable parameters may cause generation of NaN.
}

\references{
  Dr. Robert B. Abernethy, (2008) "The New Weibull Handbook, Fifth Edition" 
}

\examples{
failures<-c(90,96,30,49,82)
suspensions<-c(100,45,10)
weibull_loglike<-abremLoglike(mleframe(failures,suspensions),c( 2.26, 4900))
}

\keyword{ regression reliability }