\name{mlefit}
\alias{mlefit}

\title{Maximum likelihood regression on weibull and lognormal distributions}

\description{
 Maximum likelihood regression on weibull and lognormal distributions
}

\usage{
mlefit(x, dist="weibull")
}

\arguments{
  \item{x}{A dataframe such as generated by mleframe with column names 'left', 'right' and optionally 'qty'.
    Exact failure data (occurrences) have same time entered in both 'left' and 'right' columns.
	Suspension data has last known life[time] entered in 'left' column and -1 entered in 'right' column.
	The left(early) interval bound for left-censored data must be entered as zero. (NA is not accepted).
  }
  \item{dist}{A string defining a distribution to be fit. Implemented distributions are "weibull" (default) and "lognormal".
  (Implementations for "weibull3p" and "lognormal3p" are pending.)
  }
}

\value{
Returns a named vector with parameters of the distribution fit and goodness of fit in terms of log likelihood.
}

\references{
  Dr. Robert B. Abernethy, (2008) "The New Weibull Handbook, Fifth Edition"
  Leonard C. Johnson (1964) "The Statistical Treatment of Fatigue Experiments"  
}

\examples{
failures<-c(90,96,30,49,82)
suspensions<-c(100,45,10)
weibull_fit<-mlefit(mleframe(failures,suspensions))
}

\keyword{ regression reliability }